Add reading user password from the input and generating the password hash.
Add generating manifests for the GitOps Dashboard installation.

Split `run.go` to several files.
